requirements.txt
numpy
pandas
matplotlib
requests
cryptography
python-dotenv
openpyxl
